+++ proxy/rewriter.go
@@ -2,7 +2,6 @@ package proxy
 
 import (
 	"golang.org/x/exp/rand"
-	"net"
 	"net/http"
 )
 
@@ -13,22 +12,18 @@ type routeProvider interface {
 
 // Rewriter facilitates rewriting HTTP requests and responses according to the routes provided by a Manager.
 type Rewriter struct {
-	provider      routeProvider
-	bannedHeaders []string
+	provider   routeProvider
+	decorators []Decorator
 }
 
 // NewRewriter creates a new Rewriter backed by the given route manager.
 func NewRewriter(manager *Manager) *Rewriter {
 	return &Rewriter{
 		provider: manager,
-		bannedHeaders: []string{
-			// Variety of headers used for passing on the client IP. We don't want to pass on any rubbish clients
-			// may send in these headers. Note that we explicitly set (i.e. replace) X-Forwarded-For and
-			// X-Forwarded-Proto, so they don't need to be included here.
-			"X-Real-IP",
-			"True-Client-IP",
-			"X-Forwarded-Host",
-			"Forwarded",
+		decorators: []Decorator{
+			NewXForwardedForDecorator(),
+			NewBannedHeaderDecorator(),
+			NewUserAgentDecorator(),
 		},
 	}
 }
@@ -41,21 +36,12 @@ func (r *Rewriter) RewriteRequest(req *http.Request) {
 		return
 	}
 
-	ip, _, _ := net.SplitHostPort(req.RemoteAddr)
-	req.Header.Set("X-Forwarded-For", ip)
-	req.Header.Set("X-Forwarded-Proto", req.URL.Scheme)
+	for i := range r.decorators {
+		r.decorators[i].Decorate(req)
+	}
 
 	req.URL.Scheme = "http"
 	req.URL.Host = r.selectUpstream(route)
-
-	for i := range r.bannedHeaders {
-		req.Header.Del(r.bannedHeaders[i])
-	}
-
-	if _, ok := req.Header["User-Agent"]; !ok {
-		// explicitly disable User-Agent so it's not set to default value
-		req.Header.Set("User-Agent", "")
-	}
 }
 
 // RewriteResponse modifies the given response according to the routes provided by the Manager.