+++ proxy/rewriter_test.go
@@ -38,7 +38,7 @@ func Test_Rewriter_RewriteRequest_SetsHostToUpstream(t *testing.T) {
 
 func Test_Rewriter_RewriteRequest_SetsForwardedForHeader(t *testing.T) {
 	provider := &fakeProvider{route: &Route{Upstreams: []Upstream{{Host: "hostname:8080"}}}}
-	rewriter := &Rewriter{provider: provider}
+	rewriter := &Rewriter{provider: provider, decorators: []Decorator{NewXForwardedForDecorator()}}
 
 	u, _ := url.Parse("/foo/bar")
 	request := &http.Request{
@@ -55,7 +55,7 @@ func Test_Rewriter_RewriteRequest_SetsForwardedForHeader(t *testing.T) {
 
 func Test_Rewriter_RewriteRequest_SetsForwardedProtoHeaderIfHttps(t *testing.T) {
 	provider := &fakeProvider{route: &Route{Upstreams: []Upstream{{Host: "hostname:8080"}}}}
-	rewriter := &Rewriter{provider: provider}
+	rewriter := &Rewriter{provider: provider, decorators: []Decorator{NewXForwardedForDecorator()}}
 
 	u, _ := url.Parse("https://proxy/foo/bar")
 	request := &http.Request{
@@ -72,7 +72,7 @@ func Test_Rewriter_RewriteRequest_SetsForwardedProtoHeaderIfHttps(t *testing.T) 
 
 func Test_Rewriter_RewriteRequest_SetsForwardedProtoHeaderIfHttp(t *testing.T) {
 	provider := &fakeProvider{route: &Route{Upstreams: []Upstream{{Host: "hostname:8080"}}}}
-	rewriter := &Rewriter{provider: provider}
+	rewriter := &Rewriter{provider: provider, decorators: []Decorator{NewXForwardedForDecorator()}}
 
 	u, _ := url.Parse("http://proxy/foo/bar")
 	request := &http.Request{
@@ -89,7 +89,7 @@ func Test_Rewriter_RewriteRequest_SetsForwardedProtoHeaderIfHttp(t *testing.T) {
 
 func Test_Rewriter_RewriteRequest_ReplacesForwardedForHeader(t *testing.T) {
 	provider := &fakeProvider{route: &Route{Upstreams: []Upstream{{Host: "hostname:8080"}}}}
-	rewriter := &Rewriter{provider: provider}
+	rewriter := &Rewriter{provider: provider, decorators: []Decorator{NewXForwardedForDecorator()}}
 
 	u, _ := url.Parse("https://proxy/foo/bar")
 	request := &http.Request{
@@ -107,7 +107,7 @@ func Test_Rewriter_RewriteRequest_ReplacesForwardedForHeader(t *testing.T) {
 
 func Test_Rewriter_RewriteRequest_ReplacesForwardedProtoHeader(t *testing.T) {
 	provider := &fakeProvider{route: &Route{Upstreams: []Upstream{{Host: "hostname:8080"}}}}
-	rewriter := &Rewriter{provider: provider}
+	rewriter := &Rewriter{provider: provider, decorators: []Decorator{NewXForwardedForDecorator()}}
 
 	u, _ := url.Parse("https://proxy/foo/bar")
 	request := &http.Request{
@@ -125,7 +125,7 @@ func Test_Rewriter_RewriteRequest_ReplacesForwardedProtoHeader(t *testing.T) {
 
 func Test_Rewriter_RewriteRequest_RemovesBannedHeaders(t *testing.T) {
 	provider := &fakeProvider{route: &Route{Upstreams: []Upstream{{Host: "hostname:8080"}}}}
-	rewriter := &Rewriter{provider: provider, bannedHeaders: []string{"x-test1", "x-test2"}}
+	rewriter := &Rewriter{provider: provider, decorators: []Decorator{&bannedHeaderDecorator{[]string{"x-test1", "x-test2"}}}}
 
 	u, _ := url.Parse("https://proxy/foo/bar")
 	request := &http.Request{